requirements.txt
fastapi
uvicorn
pandas
joblib
scikit-learn
python-multipart
pydantic
pytesseract
pillow